rewrite `augur validate` interface
