+++ tests/builds/zika/Snakefile
@@ -199,7 +199,7 @@ rule export_v1:
             --output-tree {output.auspice_tree} \
             --output-meta {output.auspice_meta} \
             --output-sequence {output.auspice_seq}
-        augur validate v1 --json {output.auspice_tree} {output.auspice_meta}
+        augur validate export-v1 {output.auspice_meta} {output.auspice_tree}
         """
 
 rule export_v2:
@@ -229,7 +229,7 @@ rule export_v2:
             --output-main {output.auspice_main} \
             --output-sequence {output.auspice_seq} \
             --title {params.title:q}
-        augur validate v2 --json {output.auspice_main}
+        augur validate export-v2 {output.auspice_main}
         """
 
 rule clean:
